backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
supabase==1.2.0
drf-spectacular==0.26.5
django-cors-headers==4.3.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0 